requirements.txt
langchain==0.1.0
langchain_experimental==0.0.49
numpy==1.26.3
pandas==1.5.3
python-dotenv==1.0.0
streamlit==1.29.0
tabular==0.0.2